+++ public/client.js
@@ -1,351 +1,1 @@
-Vue.component("steam-list", {
-    props: {
-        contentTitle: String,
-        contentData: Array
-    },
-    data: function () {
-        return {
-            user: this.steamId,
-            signedIn: false
-        }
-    },
-    methods: {
-        clearData: function () {
-            this.signedIn = false;
-            this.user = "";
-            localStorage.removeItem("steamName");
-            vm.clearList();
-        },
-        getSteamId: function () {
-            console.log("click");
-            let uinput = document.getElementById("steamNum").value;
-            console.log("uinput: ", uinput);
-            if (uinput !== "") {
-                this.getSteamInfo(uinput);                
-                localStorage.setItem("steamName", uinput);
-            }
-        },
-        getSteamInfo: function (user) {
-            $.ajax({
-                type: "POST",
-                url: "/steam",
-                headers: { "username": user },
-                success: function (data) {
-                    vm.setSteamList(data);
-                    this.signedIn = true;
-                    $("#steam").css({ "color": "white" });
-                },
-                error: function (err) {
-                    console.log(err.statusCode);
-                }
-            });
-        }
-    },
-    template: `
-            <div class="content">
-            <h3>Steam table goes here...</h3>
-            <input type="text" id="steamNum" />
-            <button id="#getsteam" @click="getSteamId">Get Steam data</button>
-            <div v-if="this.signedIn === true">
-                <section id="gamelist">
-                    <template v-for="item in contentData[0]">
-                        <p>{{item}}</p>
-                  </template>
-                </section>
-                <section id="friendList">
-                    <template v-for="item in contentData[1]">
-                        <p>{{item}}</p>
-                    </template>
-                </section>
-                </div>
-            </div>
-        `
-});
-
-Vue.component("steam-item", {
-    props: ["type"],
-    template: `<p>Testing steam-item {{type}} component</p>`
-});
-
-
-Vue.component("twitch-list", {
-    props: {
-        contentTitle: String,
-        contentData: Array,
-        contentType: String,
-        clearList: Function,
-        getStreamList: Function,
-        twitchName: String
-    },
-    data: function () {
-        return {
-            user: this.twitchName,
-            signedIn: false
-        }
-    },
-    methods: {
-        getName: function () {
-            let input = document.getElementById("username");
-            if (input.value !== "") {
-                user = input.value.toLowerCase();
-                vm.getStreamList(user);
-                this.signedIn = true;
-                localStorage.setItem("twitchName", user);
-            }
-        },
-        clearData: function () {
-            this.signedIn = false;
-            this.user = "";
-            localStorage.removeItem("twitchName");
-            vm.clearList();
-        }
-    },
-    mounted() {
-        let x = localStorage.getItem("twitchName");
-        if (x) {
-            this.signedIn = true;
-            this.user = x;
-        }
-    },
-    template: `
-        <div class="content">
-            <section v-if="this.signedIn === false">
-                <label for="username">Enter Twitch.tv username</label>
-                <input type="text" v-model=user id="username"/>
-                <button class="btn-filter" id="twitch-auth" @click=getName()>Get follow list</button></button>
-            </section>
-            <section v-if="this.signedIn === true">
-                <span>Signed in as {{user}}</span>
-                <button class="btn-filter" id="twitch-signout" style="display: block;" @click=clearData()>Clear Twitch list</button>
-            </section>
-
-        <table v-if="this.signedIn === true">
-            <caption class="hidden" aria-hidden="false">{{contentTitle}}</caption>
-                <thead>
-                    <tr>
-                        <th scope="col">Channel</th>
-                        <th scope="col">Name</th>
-                        <th scope="col">Description</th>
-                    </tr>
-                </thead>
-                <tbody>                    
-                    <template v-for="item in contentData">
-                        <twitch-result
-                            :logo="item.logo"
-                            :name="item.display_name"
-                            :status="item.title"
-                            :url="item.url"                                                       
-                        ></twitch-result> 
-                    </template>                       
-                </tbody>
-            </table>
-        </div>
-    `
-})
-
-Vue.component("youtube-list", {
-    props: {
-        contentTitle: String,
-        contentData: Array,
-        contentType: String
-    },
-    methods: {
-        snip: function (string) {
-            if (string.length > 100) { return string.substr(0, 200) + "..."; }
-            else { return string; }
-        }
-    },
-    template: `
-        <div class="content">
-            <button class="btn-filter" id="authorize-button" style="display: block;">Authorize Y</button>
-            <button class="btn-filter" id="signout-button" style="display: block;">Sign out of YouTube</button>
-            <table v-if="this.contentData.length > 0">
-                <caption class="hidden" aria-hidden="false">{{contentTitle}}</caption>
-            <thead>
-                <tr>
-                    <th scope="col">Channel</th>
-                    <th scope="col">Name</th>
-                    <th scope="col">Description</th>
-                </tr>
-            </thead>
-            <tbody>   
-                <template v-for="item in contentData">
-                    <template v-if="item.snippet && item.contentDetails.newItemCount > 0">
-                        <youtube-result
-                            :desc="snip(item.snippet.description)" 
-                            :logo="item.snippet.thumbnails.default.url" 
-                            :name="item.snippet.title" 
-                            :newItem="item.contentDetails.newItemCount"                                                                               
-                            :url="item.snippet.resourceId.channelId"                                                       
-                        ></youtube-result> 
-                    </template>
-                </template>            
-                </tbody>
-            </table>
-        </div>
-            `
-})
-
-Vue.component("twitch-result", {
-    props: ["game", "logo", "name", "status", "url"],   //All strings
-    template: `<tr class="row">
-                <th scope="row"><img :src="logo" :alt="name + ' stream logo'"></th>
-                <td><span class="names"><a :href="url"> {{ name }} </a></span></td>
-                <td><span> {{ game }} </span> </td>
-                <td><span class="sm-hide"> <a :href="url"> {{ status }} </a></span></td>
-                </tr>`,
-})
-
-Vue.component("youtube-result", {
-    props: {
-        desc: {
-            type: String,
-            default: "No description given"
-        },
-        logo: String,
-        name: String,
-        newItem: {
-            type: [Number, String],
-            default: 0
-        },
-        url: String
-    },
-    template: `<tr class="row">
-                <th scope="row"><img :src="logo" :alt="name + ' channel logo'"><span> {{newItem}} </span></th>
-                <td><span class="names"><a :href="'https://youtube.com/channel/' + url + '/videos'"> {{ name }} </a></span></td>
-                <td><span class="sm-hide"> <a :href="'https://youtube.com/channel/' + url + '/videos'"> {{ desc }} </a></span></td>
-                </tr>`,
-})
-
-
-var vm = new Vue({
-    el: "#vue-app",
-    data: {
-        twitchResults: [],
-        ytResults: [],
-        steamResults: [],
-        twitchName: "",
-        steamId: ""
-    },
-    methods: {
-        setUser: function (user) {
-            this.twitchName = user;
-        },
-        getStreamList: function (user) {
-            $.ajax({
-                type: "POST",
-                url: "/streams",
-                headers: { "username": user },
-                success: function (data) {
-                    vm.setStreamList(data);
-                    $("#games").css({ "color": "#4B367C" });
-                },
-                error: function (err) {
-                    console.error(err);
-                }
-            });
-        },
-        setStreamList: function (data) {
-            for (var i = 0; i < data[0].length; i++) {
-                data[0][i]["display_name"] = data[1][i]["display_name"];
-                data[0][i]["logo"] = data[1][i]["profile_image_url"];
-            }
-            this.twitchResults = data[0];
-        },
-        setSteamList: function (data) {
-            this.steamResults = data;
-            console.log("steamResults: ", data);
-        },
-        getVideoList: function (user) {
-            //Thanks to Google API code, the videos pretty much get themselves.
-        },
-        setVideoList: function (data) {
-            this.ytResults = data;
-            $("#videos").css({ "color": "red" });
-
-        },
-        clearList: function () {
-            this.twitchResults = [];
-        }
-    },
-    mounted() {
-        let x = localStorage.getItem("twitchName");
-        if (x) {
-            this.setUser(x, "twitchName");
-            this.getStreamList(this.twitchName);
-        }
-    }
-});
-
-/*
-The following code is from the YouTube Data API quickstart guide, with some slight modifications.
-See https://developers.google.com/youtube/v3/quickstart/js.
-*/
-
-var CLIENT_ID = "372774319049-pk9d85udr75rlqcluuq12apdeqtnk8go.apps.googleusercontent.com";
-var DISCOVERY_DOCS = ["https://www.googleapis.com/discovery/v1/apis/youtube/v3/rest"];
-var SCOPES = 'https://www.googleapis.com/auth/youtube.readonly';
-
-var authorizeButton = document.getElementById('authorize-button');
-var signoutButton = document.getElementById('signout-button');
-
-/**
- *  On load, called to load the auth2 library and API client library.
- */
-function handleClientLoad() {
-    gapi.load('client:auth2', initClient);
-}
-
-/**
- *  Initializes the API client library and sets up sign-in state
- *  listeners.
- */
-function initClient() {
-    gapi.client.init({
-        discoveryDocs: DISCOVERY_DOCS,
-        clientId: CLIENT_ID,
-        scope: SCOPES
-    }).then(function () {
-        // Listen for sign-in state changes.
-        gapi.auth2.getAuthInstance().isSignedIn.listen(updateSigninStatus);
-
-        // Handle the initial sign-in state.
-        updateSigninStatus(gapi.auth2.getAuthInstance().isSignedIn.get());
-        authorizeButton.onclick = handleAuthClick;
-        signoutButton.onclick = handleSignoutClick;
-    });
-}
-
-/**
- *  Called when the signed in status changes, to update the UI
- *  appropriately. After a sign-in, the API is called.
- */
-function updateSigninStatus(isSignedIn) {
-    if (isSignedIn) {
-        authorizeButton.style.display = 'none';
-        signoutButton.style.display = 'block';
-        getSubscriptions();
-    } else {
-        authorizeButton.style.display = 'block';
-        signoutButton.style.display = 'none';
-        vm.ytResults = [{}];
-    }
-}
-
-function handleAuthClick(event) {
-    gapi.auth2.getAuthInstance().signIn();
-}
-
-function handleSignoutClick(event) {
-    gapi.auth2.getAuthInstance().signOut();
-}
-
-function getSubscriptions() {
-    gapi.client.youtube.subscriptions.list({
-        'part': 'snippet,contentDetails',
-        'mine': 'true',
-        'maxResults': 20
-    }).then(function (response) {
-        vm.setVideoList(response.result.items);
-    });
-}
+"use strict";function handleClientLoad(){gapi.load("client:auth2",initClient)}function initClient(){gapi.client.init({discoveryDocs:DISCOVERY_DOCS,clientId:CLIENT_ID,scope:SCOPES}).then(function(){gapi.auth2.getAuthInstance().isSignedIn.listen(updateSigninStatus),updateSigninStatus(gapi.auth2.getAuthInstance().isSignedIn.get()),authorizeButton.onclick=handleAuthClick,signoutButton.onclick=handleSignoutClick})}function updateSigninStatus(t){t?(authorizeButton.style.display="none",signoutButton.style.display="block",getSubscriptions()):(authorizeButton.style.display="block",signoutButton.style.display="none",vm.ytResults=[{}])}function handleAuthClick(t){gapi.auth2.getAuthInstance().signIn()}function handleSignoutClick(t){gapi.auth2.getAuthInstance().signOut()}function getSubscriptions(){gapi.client.youtube.subscriptions.list({part:"snippet,contentDetails",mine:"true",maxResults:20}).then(function(t){vm.setVideoList(t.result.items)})}var vm=new Vue({el:"#vue-app",data:{twitchResults:[],ytResults:[],steamResults:[],twitchName:"",steamId:""},methods:{setUser:function(t){this.twitchName=t},getStreamList:function(t){$.ajax({type:"POST",url:"/streams",headers:{username:t},success:function(t){vm.setStreamList(t),$("#games").css({color:"#4B367C"})},error:function(t){console.error(t)}})},setStreamList:function(t){for(var e=0;e<t[0].length;e++)t[0][e].display_name=t[1][e].display_name,t[0][e].logo=t[1][e].profile_image_url;this.twitchResults=t[0]},setSteamList:function(t){this.steamResults=t,console.log("steamResults: ",t)},getVideoList:function(t){},setVideoList:function(t){this.ytResults=t,$("#videos").css({color:"red"})},clearList:function(){this.twitchResults=[]}},mounted:function(){var t=localStorage.getItem("twitchName");t&&(this.setUser(t,"twitchName"),this.getStreamList(this.twitchName))}}),CLIENT_ID="372774319049-pk9d85udr75rlqcluuq12apdeqtnk8go.apps.googleusercontent.com",DISCOVERY_DOCS=["https://www.googleapis.com/discovery/v1/apis/youtube/v3/rest"],SCOPES="https://www.googleapis.com/auth/youtube.readonly",authorizeButton=document.getElementById("authorize-button"),signoutButton=document.getElementById("signout-button");
