+++ public/index.html
@@ -10,7 +10,7 @@
     <link href="https://fonts.googleapis.com/css?family=Cabin|Open+Sans" rel="stylesheet">
     <script src="https://use.fontawesome.com/59c976fd61.js"></script>
     <script src="https://unpkg.com/vue@2.5.2/dist/vue.js"></script>
-    <link rel="stylesheet" href="/public/style.css">
+    <link rel="stylesheet" href="/style.css">
 </head>
 
 <body>
@@ -42,7 +42,7 @@
     <script src="https://code.jquery.com/jquery-3.1.1.min.js" integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8=" crossorigin="anonymous"></script>
     <script async defer src="https://apis.google.com/js/api.js" onload="this.onload=function(){};handleClientLoad()" onreadystatechange="if (this.readyState === 'complete') this.onload()">
     </script>
-    <script src="/public/client.js"></script>
+    <script src="/client.js"></script>
 </body>
 
 </html>
