+++ public/routes.js
@@ -1,0 +1,1 @@
+"use strict";var express=require("express"),async=require("async"),bodyParser=require("body-parser"),https=require("https"),request=require("request"),rp=require("request-promise"),router=express.Router(),KEY_T=process.env.KEY_T,KEY_Y=process.env.KEY_Y,KEY_S=process.env.KEY_S,T_DATA=[],tCode="";router.get("/",function(e,t){t.sendFile(__dirname+"/public/index.html")}),router.post("/steam",function(e,t,r){var s=e.headers.username,n=new Promise(function(e,t){rp({uri:"http://api.steampowered.com/ISteamUser/GetFriendList/v0001/?key="+KEY_S+"&steamid="+s+"&relationship=friend",json:!0}).then(function(t){var r=t,n=[];for(var i in r.friendslist.friends)n.push(r.friendslist.friends[i].steamid);n=n.join(","),rp({uri:"http://api.steampowered.com/ISteamUser/GetPlayerSummaries/v0002/?key="+KEY_S+"&steamids="+n,json:!0}).then(function(t){var r=[];r.push(t),rp({uri:"http://api.steampowered.com/IPlayerService/GetOwnedGames/v0001/?key="+KEY_S+"&include_appinfo=1&steamid="+s+"&format=json",json:!0}).then(function(t){r.push(t),e(r)}).catch(function(e){console.log(e)})}).catch(function(e){console.log(e)})}).catch(function(e){console.log(e)})});Promise.resolve(n).then(function(e){t.send(e)})}),router.post("/streams",function(e,t,r){var s=e.headers.username,n=new Promise(function(e,t){rp({uri:"https://api.twitch.tv/helix/users?login="+s,headers:{"Client-ID":KEY_T},json:!0}).then(function(t){var r=t.data[0].id;rp({uri:"https://api.twitch.tv/helix/users/follows?first=100&from_id="+r,headers:{"Client-ID":KEY_T},json:!0}).then(function(t){var r=[];for(var s in t.data)r.push(t.data[s].to_id);var n=r.join("&user_id=");rp({uri:"https://api.twitch.tv/helix/streams?first=100&user_id="+n,headers:{"Client-ID":KEY_T},json:!0}).then(function(t){var r=[],s=[];r.push(t.data);for(var n in t.data)"live"===t.data[n].type&&s.push(t.data[n].user_id);var i={uri:"https://api.twitch.tv/helix/users?id="+s.join("&id="),headers:{"Client-ID":KEY_T},json:!0};rp(i).then(function(t){r.push(t.data),e(r)}).catch(function(e){console.log(e)})}).catch(function(e){console.log(e)})}).catch(function(e){console.log(e)})}).catch(function(e){console.log(e)})});Promise.resolve(n).then(function(e){t.send(e)})}),module.exports=router;
